+++ client/src/test/java/com/microsoft/durabletask/IntegrationTestBase.java
@@ -69,5 +69,10 @@ public class IntegrationTestBase {
             });
             return this;
         }
+
+        public TestDurableTaskWorkerBuilder useVersioning(DurableTaskGrpcWorkerVersioningOptions options) {
+            this.innerBuilder.useVersioning(options);
+            return this;
+        }
     }
 }