+++ client/src/test/java/com/microsoft/durabletask/IntegrationTests.java
@@ -42,8 +42,9 @@ public class IntegrationTests extends IntegrationTestBase {
     // Before whole test suite, delete the task hub
     @BeforeEach
     private void startUp() {
-        DurableTaskClient client = new DurableTaskGrpcClientBuilder().build();
-        client.deleteTaskHub();
+        try (DurableTaskClient client = new DurableTaskGrpcClientBuilder().build()) {
+            client.deleteTaskHub();
+        }
     }
 
     @AfterEach
@@ -53,7 +54,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void emptyOrchestration() throws TimeoutException {
         final String orchestratorName = "EmptyOrchestration";
         final String input = "Hello " + Instant.now();
@@ -76,7 +77,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void singleTimer() throws IOException, TimeoutException {
         final String orchestratorName = "SingleTimer";
         final Duration delay = Duration.ofSeconds(3);
@@ -99,7 +100,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void longTimer() throws TimeoutException {
         final String orchestratorName = "LongTimer";
         final Duration delay = Duration.ofSeconds(7);
@@ -143,7 +144,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void longTimerNonblocking() throws TimeoutException {
         final String orchestratorName = "ActivityAnyOf";
         final String externalEventActivityName = "externalEvent";
@@ -181,7 +182,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void longTimerNonblockingNoExternal() throws TimeoutException {
         final String orchestratorName = "ActivityAnyOf";
         final String externalEventActivityName = "externalEvent";
@@ -218,7 +219,7 @@ public class IntegrationTests extends IntegrationTestBase {
     }
 
 
-    @RetryingTest
+    @Test
     void longTimeStampTimer() throws TimeoutException {
         final String orchestratorName = "LongTimeStampTimer";
         final Duration delay = Duration.ofSeconds(7);
@@ -252,7 +253,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void singleTimeStampTimer() throws IOException, TimeoutException {
         final String orchestratorName = "SingleTimeStampTimer";
         final Duration delay = Duration.ofSeconds(3);
@@ -276,7 +277,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void isReplaying() throws IOException, InterruptedException, TimeoutException {
         final String orchestratorName = "SingleTimer";
         DurableTaskGrpcWorker worker = this.createWorkerBuilder()
@@ -312,7 +313,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void singleActivity() throws IOException, InterruptedException, TimeoutException {
         final String orchestratorName = "SingleActivity";
         final String activityName = "Echo";
@@ -344,7 +345,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void currentDateTimeUtc() throws IOException, TimeoutException {
         final String orchestratorName = "CurrentDateTimeUtc";
         final String echoActivityName = "Echo";
@@ -383,7 +384,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void activityChain() throws IOException, TimeoutException {
         final String orchestratorName = "ActivityChain";
         final String plusOneActivityName = "PlusOne";
@@ -410,7 +411,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void subOrchestration() throws TimeoutException {
         final String orchestratorName = "SubOrchestration";
         DurableTaskGrpcWorker worker = this.createWorkerBuilder().addOrchestrator(orchestratorName, ctx -> {
@@ -431,7 +432,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void continueAsNew() throws TimeoutException {
         final String orchestratorName = "continueAsNew";
         final Duration delay = Duration.ofSeconds(0);
@@ -454,7 +455,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void continueAsNewWithExternalEvents() throws TimeoutException, InterruptedException{
         final String orchestratorName = "continueAsNewWithExternalEvents";
         final String eventName = "MyEvent";
@@ -485,7 +486,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void termination() throws TimeoutException {
         final String orchestratorName = "Termination";
         final Duration delay = Duration.ofSeconds(3);
@@ -507,7 +508,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingParameterizedTest
+    @ParameterizedTest
     @ValueSource(booleans = {true, false})
     void restartOrchestrationWithNewInstanceId(boolean restartWithNewInstanceId) throws TimeoutException {
         final String orchestratorName = "restart";
@@ -534,7 +535,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void restartOrchestrationThrowsException() {
         final String orchestratorName = "restart";
         final Duration delay = Duration.ofSeconds(3);
@@ -596,7 +597,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void terminateSuspendOrchestration() throws TimeoutException, InterruptedException {
         final String orchestratorName = "suspendResume";
         final String eventName = "MyEvent";
@@ -622,7 +623,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void activityFanOut() throws IOException, TimeoutException {
         final String orchestratorName = "ActivityFanOut";
         final String activityName = "ToString";
@@ -664,7 +665,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void externalEvents() throws IOException, TimeoutException {
         final String orchestratorName = "ExternalEvents";
         final String eventName = "MyEvent";
@@ -703,7 +704,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingParameterizedTest
+    @ParameterizedTest
     @ValueSource(booleans = {true, false})
     void externalEventsWithTimeouts(boolean raiseEvent) throws IOException, TimeoutException {
         final String orchestratorName = "ExternalEventsWithTimeouts";
@@ -742,7 +743,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void setCustomStatus() throws TimeoutException {
         final String orchestratorName = "SetCustomStatus";
 
@@ -775,7 +776,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void clearCustomStatus() throws TimeoutException {
         final String orchestratorName = "ClearCustomStatus";
 
@@ -805,7 +806,7 @@ public class IntegrationTests extends IntegrationTestBase {
     }
 
     // due to clock drift, client/worker and sidecar time are not exactly synchronized, this test needs to accommodate for client vs backend timestamps difference
-    @RetryingTest
+    @Test
     void multiInstanceQuery() throws TimeoutException{
         final String plusOne = "plusOne";
         final String waitForEvent = "waitForEvent";
@@ -986,7 +987,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void purgeInstanceId() throws TimeoutException {
         final String orchestratorName = "PurgeInstance";
         final String plusOneActivityName = "PlusOne";
@@ -1017,7 +1018,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void purgeInstanceFilter() throws TimeoutException {
         final String orchestratorName = "PurgeInstance";
         final String plusOne = "PlusOne";
@@ -1114,7 +1115,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void purgeInstanceFilterTimeout() throws TimeoutException {
         final String orchestratorName = "PurgeInstance";
         final String plusOne = "PlusOne";
@@ -1171,7 +1172,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void waitForInstanceStartThrowsException() {
         final String orchestratorName = "orchestratorName";
 
@@ -1193,7 +1194,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void waitForInstanceCompletionThrowsException() {
         final String orchestratorName = "orchestratorName";
         final String plusOneActivityName = "PlusOne";
@@ -1223,7 +1224,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void activityFanOutWithException() throws TimeoutException {
         final String orchestratorName = "ActivityFanOut";
         final String activityName = "Divide";
@@ -1280,7 +1281,7 @@ public class IntegrationTests extends IntegrationTestBase {
                 expectedExceptionMessage);
     }
 
-    @RetryingTest
+    @Test
     void thenApply() throws IOException, InterruptedException, TimeoutException {
         final String orchestratorName = "thenApplyActivity";
         final String activityName = "Echo";
@@ -1313,7 +1314,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void externalEventThenAccept() throws InterruptedException, TimeoutException {
         final String orchestratorName = "continueAsNewWithExternalEvents";
         final String eventName = "MyEvent";
@@ -1347,7 +1348,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void activityAllOf() throws IOException, TimeoutException {
         final String orchestratorName = "ActivityAllOf";
         final String activityName = "ToString";
@@ -1406,7 +1407,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void activityAllOfException() throws IOException, TimeoutException {
         final String orchestratorName = "ActivityAllOf";
         final String activityName = "ToString";
@@ -1468,7 +1469,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     void activityAnyOf() throws IOException, TimeoutException {
         final String orchestratorName = "ActivityAnyOf";
         final String activityName = "ToString";
@@ -1517,7 +1518,7 @@ public class IntegrationTests extends IntegrationTestBase {
         }
     }
 
-    @RetryingTest
+    @Test
     public void newUUIDTest() {
         String orchestratorName = "test-new-uuid";
         String echoActivityName = "Echo";
@@ -1560,6 +1561,183 @@ public class IntegrationTests extends IntegrationTestBase {
         } catch (TimeoutException e) {
             throw new RuntimeException(e);
         }
+    }
 
+    @Test
+    public void defaultVersionPassedThroughToContext() {
+        final String orchestratorName = "VersionOrchestration";
+        final String activityName = "SayVersion";
+        final String defaultVersion = "1.0";
+        DurableTaskGrpcWorker worker = this.createWorkerBuilder()
+            .addOrchestrator(orchestratorName, ctx -> {
+                String version = ctx.getVersion();
+                String output = ctx.callActivity(activityName, version, String.class).await();
+                ctx.complete(output);
+            })
+            .addActivity(activityName, ctx -> {
+                return String.format("Version: %s", ctx.getInput(String.class));
+            })
+            .buildAndStart();
+
+        DurableTaskClient client = new DurableTaskGrpcClientBuilder()
+            .defaultVersion(defaultVersion)
+            .build();
+        try (worker; client) {
+            String instanceId = client.scheduleNewOrchestrationInstance(orchestratorName);
+            OrchestrationMetadata instance = client.waitForInstanceCompletion(
+                instanceId,
+                defaultTimeout,
+                true);
+
+            assertNotNull(instance);
+            assertEquals(OrchestrationRuntimeStatus.COMPLETED, instance.getRuntimeStatus());
+            String output = instance.readOutputAs(String.class);
+            String expected = String.format("Version: %s", defaultVersion);
+            assertEquals(expected, output);
+        } catch (TimeoutException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Test
+    public void optionVersionOverridesDefault() {
+        final String orchestratorName = "VersionOrchestration";
+        final String activityName = "SayVersion";
+        final String defaultVersion = "1.0";
+        final String overrideVersion = "2.0";
+        DurableTaskGrpcWorker worker = this.createWorkerBuilder()
+            .addOrchestrator(orchestratorName, ctx -> {
+                String version = ctx.getVersion();
+                String output = ctx.callActivity(activityName, version, String.class).await();
+                ctx.complete(output);
+            })
+            .addActivity(activityName, ctx -> {
+                return String.format("Version: %s", ctx.getInput(String.class));
+            })
+            .buildAndStart();
+
+        DurableTaskClient client = new DurableTaskGrpcClientBuilder()
+            .defaultVersion(defaultVersion)
+            .build();
+        try (worker; client) {
+            String instanceId = client.scheduleNewOrchestrationInstance(orchestratorName, new NewOrchestrationInstanceOptions()
+                .setVersion(overrideVersion));
+            OrchestrationMetadata instance = client.waitForInstanceCompletion(
+                instanceId,
+                defaultTimeout,
+                true);
+
+            assertNotNull(instance);
+            assertEquals(OrchestrationRuntimeStatus.COMPLETED, instance.getRuntimeStatus());
+            String output = instance.readOutputAs(String.class);
+            String expected = String.format("Version: %s", overrideVersion);
+            assertEquals(expected, output);
+        } catch (TimeoutException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"", "0.9", "1.0", "1.1"})
+    void orchestrationVersionCurrentOrOlderMatchStrategy(String orchestrationVersion) {
+        final String orchestratorName = "VersionedOrchestrationFailTest";
+        final String activityName = "SayVersion";
+        final DurableTaskGrpcWorkerVersioningOptions versioningOptions = new DurableTaskGrpcWorkerVersioningOptions(
+            "1.0", 
+            "1.0", 
+            DurableTaskGrpcWorkerVersioningOptions.VersionMatchStrategy.CURRENTOROLDER, 
+            DurableTaskGrpcWorkerVersioningOptions.VersionFailureStrategy.FAIL);;
+
+        DurableTaskGrpcWorker worker = this.createWorkerBuilder()
+            .addOrchestrator(orchestratorName, ctx -> {
+                String version = ctx.getVersion();
+                String output = ctx.callActivity(activityName, version, String.class).await();
+                ctx.complete(output);
+            })
+            .addActivity(activityName, ctx -> String.format("Version: %s", ctx.getInput(String.class)))
+            .useVersioning(versioningOptions)
+            .buildAndStart();
+
+        DurableTaskClient client = new DurableTaskGrpcClientBuilder().build();
+        try (worker; client) {
+            String instanceId = client.scheduleNewOrchestrationInstance(
+                orchestratorName,
+                new NewOrchestrationInstanceOptions().setVersion(orchestrationVersion)
+            );
+            if (orchestrationVersion.equals("1.1")) {
+                OrchestrationMetadata instance = client.waitForInstanceCompletion(
+                    instanceId,
+                    defaultTimeout,
+                    true
+                );
+                assertNotNull(instance);
+                assertEquals(OrchestrationRuntimeStatus.FAILED, instance.getRuntimeStatus());
+            } else {
+                OrchestrationMetadata instance = client.waitForInstanceCompletion(
+                    instanceId,
+                    defaultTimeout,
+                    true
+                );
+                assertNotNull(instance);
+                assertEquals(OrchestrationRuntimeStatus.COMPLETED, instance.getRuntimeStatus());
+                String output = instance.readOutputAs(String.class);
+                String expected = String.format("Version: %s", orchestrationVersion);
+                assertEquals(expected, output);
+            }
+        } catch (TimeoutException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"", "0.9", "1.0", "1.1"})
+    void orchestrationVersionStrictMatchStrategy(String orchestrationVersion) {
+        final String orchestratorName = "VersionedOrchestrationFailTestStrict";
+        final String activityName = "SayVersion";
+        final DurableTaskGrpcWorkerVersioningOptions versioningOptions = new DurableTaskGrpcWorkerVersioningOptions(
+            "1.0",
+            "1.0",
+            DurableTaskGrpcWorkerVersioningOptions.VersionMatchStrategy.STRICT,
+            DurableTaskGrpcWorkerVersioningOptions.VersionFailureStrategy.FAIL);
+
+        DurableTaskGrpcWorker worker = this.createWorkerBuilder()
+            .addOrchestrator(orchestratorName, ctx -> {
+                String version = ctx.getVersion();
+                String output = ctx.callActivity(activityName, version, String.class).await();
+                ctx.complete(output);
+            })
+            .addActivity(activityName, ctx -> String.format("Version: %s", ctx.getInput(String.class)))
+            .useVersioning(versioningOptions)
+            .buildAndStart();
+
+        DurableTaskClient client = new DurableTaskGrpcClientBuilder().build();
+        try (worker; client) {
+            String instanceId = client.scheduleNewOrchestrationInstance(
+                orchestratorName,
+                new NewOrchestrationInstanceOptions().setVersion(orchestrationVersion)
+            );
+            if (orchestrationVersion.equals("1.0")) {
+                OrchestrationMetadata instance = client.waitForInstanceCompletion(
+                    instanceId,
+                    defaultTimeout,
+                    true
+                );
+                assertNotNull(instance);
+                assertEquals(OrchestrationRuntimeStatus.COMPLETED, instance.getRuntimeStatus());
+                String output = instance.readOutputAs(String.class);
+                String expected = String.format("Version: %s", orchestrationVersion);
+                assertEquals(expected, output);
+            } else {
+                OrchestrationMetadata instance = client.waitForInstanceCompletion(
+                    instanceId,
+                    defaultTimeout,
+                    true
+                );
+                assertNotNull(instance);
+                assertEquals(OrchestrationRuntimeStatus.FAILED, instance.getRuntimeStatus());
+            }
+        } catch (TimeoutException e) {
+            throw new RuntimeException(e);
+        }
     }
 }